requirements.txt
customtkinter==5.2.0
CTkMessagebox==2.0
tkinterdnd2==0.3.0
Pillow==10.0.0
pypdf>=4.0.0
python-docx==0.8.11
python-pptx==0.6.21
pandas==2.0.0
openpyxl==3.1.0
markdown==3.4.0
beautifulsoup4==4.12.0
chardet==5.2.0
httpx>=0.27.0
openai==1.0.0
anthropic==0.18.0
groq==0.4.0
langdetect==1.0.9
nltk==3.8.0
reportlab>=4.0.0
pydantic==2.0.0
pyinstaller==6.0.0
